requirements.txt
flask
requests
jinja2
gunicorn
python-dotenv
